[code-format] Switch between formatCode and formatEntireFile
This addresses the issue from
https://github.com/facebook/nuclide/commit/73b8c7346aa419bc4f05216ab578aeffb3f617a2#commitcomment-15965354
where we check for the presence of either formatCode or formatEntireFile
without taking into consideration whether there's a selected region.

The future logic for this should probably change, but this current
behavior is backward-compatible.

Test plan:

Try it on a formatter that provides both options.