Restrict max workers to something sane
When the number of CPUs on a machine is quite large, flow spawns enough processes to saturate IO and starve everything else on the machine. This change makes it so that on machines with a very large number of cores, flow plays more nicely with other things (arc, hh, etc) happening concurrently. This is especially impactful when you do a rebase that touches a lot of files, and try to run a lot of unit tests or something shortly after.

See https://www.desmos.com/calculator/or5ppybxia for a graph of current versus new algos for max workers.