Swap navigation forward/backward keys on PCs
Windows and Linux used 'Ctrl-<' to go forwards and 'Ctrl->' to
go backwards in history.  This is opposite the expected direction.
Swap these two key definitions.

The Mac/OS X key definitions were correct, but the documentation also
had them backwards.  Swap the key definition descriptions in the
documentation.

A remaining issue appears to exist which blocks Ctrl-< since it is
bound to a different action by Atom-Core.  This change doesn't correct
that because it's a different issue and it's not obvious how it should
be fixed in Nuclide.